+++ ct/alpine-traefik.sh
@@ -38,4 +38,4 @@ description
 msg_ok "Completed Successfully!\n"
 echo -e "${CREATING}${GN}${APP} setup has been successfully initialized!${CL}"
 echo -e "${INFO}${YW} WebUI Access (if configured) - using the following URL:${CL}"
-echo -e "${TAB}${GATEWAY}${BGN}https://${IP}:8080/dashboard${CL}"
+echo -e "${TAB}${GATEWAY}${BGN}http://${IP}:8080/dashboard${CL}"