+++ ct/homeassistant.sh
@@ -95,7 +95,7 @@ WantedBy=default.target" >$service_path
 
     msg_ok "Completed Successfully!\n"
     echo -e "FileBrowser should be reachable by going to the following URL.
-         ${BL}https://$IP:8080${CL}   admin|helper-scripts.com\n"
+         ${BL}http://$IP:8080${CL}   admin|helper-scripts.com\n"
     exit
   fi
 }
@@ -107,5 +107,5 @@ description
 msg_ok "Completed Successfully!\n"
 echo -e "${CREATING}${GN}${APP} setup has been successfully initialized!${CL}"
 echo -e "${INFO}${YW} Access it using the following URL:${CL}"
-echo -e "${TAB}${GATEWAY}${BGN}HA: https://${IP}:8123${CL}"
+echo -e "${TAB}${GATEWAY}${BGN}HA: http://${IP}:8123${CL}"
 echo -e "${TAB}${GATEWAY}${BGN}Portainer: https://${IP}:9443${CL}"