+++ CHANGELOG.md
@@ -6359,7 +6359,7 @@ Exercise vigilance regarding copycat or coat-tailing sites that seek to exploit 
 
 ### Changed
 
-- **The site can now be accessed through a more memorable URL, which is [helper-scripts.com](https://helper-scripts.com).**
+- **The site can now be accessed through a more memorable URL, which is [helper-scripts.com](http://helper-scripts.com).**
 
 ## 2023-04-23